src/sgame, src/enemyc.c, src/combat.c: Using size_t instead of unsigned int
src/masmenu.c: Voiding out unused parameter

src/kq.c: Using size_t instead of unsigned int, moved redundant multiplication calls into local variable, cast values that should be treated as unsigned integers to size_t

src/intrface.c: Added comments, fixed/added parenthesis around ?: pairs

src/eqpmenu.c: Explicitely cast signed value to unsigned in function that requires unsigned types, using size_t instead of unsigned short

src/draw.c, include/platform.h, enums.h: Typedef of enumeration to prevent redundant usage later

+++ KQ/src/enemyc.c
@@ -743,7 +743,7 @@ static s_fighter *make_enemy (int who, s_fighter * en)
  */
 int select_encounter (int en, int etid)
 {
-   int i, p, j;
+   size_t i, p, j;
    int stop = 0, where = 0, entry = -1;
 
    while (!stop) {
